+++ src/games/_base.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-classes-per-file */
 import { APGamesInformation, Variant } from '../schemas/gameinfo';
 import { APRenderRep, Glyph } from "@abstractplay/renderer/src/schemas/schema";
 import { APMoveResult } from '../schemas/moveresults';
@@ -182,7 +183,7 @@ export abstract class GameBase  {
     public abstract results: Array<APMoveResult>;
     public abstract variants: string[];
 
-    public abstract move(move: string): GameBase;
+    public abstract move(move: string, partial?: boolean): GameBase;
     public abstract render(perspective?: any): APRenderRep;
     public abstract state(): IAPGameState;
     public abstract load(idx: number): GameBase;
@@ -199,6 +200,10 @@ export abstract class GameBase  {
 
     public draw(): GameBase {
         return this.eog(-1, "draw", {type: "drawagreed"});
+    }
+
+    protected specialMove(move: string): boolean {
+        return move === "resign" || move === "draw" || move === "timeout";
     }
 
     private eog(player: number, move: string, result: APMoveResult): GameBase {
@@ -612,5 +617,12 @@ export abstract class GameBase  {
 
     public serialize(): string {
         return JSON.stringify(this.state(), replacer);
+    }
+}
+
+export abstract class GameBaseSimultaneous extends GameBase {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    public isEliminated(id: number): boolean {
+        return false;
     }
 }